Handle terminal resizes.
The KEY_RESIZE feature of ncurses is not used since it is not possible
to detect if a KEY_RESIZE was queued by selecting on stdin. Hence, tinit
now installs a SIGWINCH handler which will set the winchg variable to 1.

I rely on the fact that select will be interrupted by the signal which
pops after a terminal resize to be able to redraw the screen
instantaneously.

tresize does all the job of resizing the three used curses windows.