Display joined channels in status bar.
The status bar was not useful, now it displays the list of joined
channels. The function tdrawbar will redraw the whole bar (no clever
refreshing here, it does not worth the complexity). Tdrawbar tries to
keep the current channel in the middle of the bar to maximize the
context on both sides.