Fix uparse to deal explicitely with empty lines.
Uparse accessed undefined elements if the input line was empty. This was
harmless because the default branch of the switch led the control flow
to the right place, however it could appear a bit obscure. Adding an
extra test will not hurt performance and make the code easier to read.